Omegam split rescaled generic linear and generic nonlinear boost in split_cosmology
